+++ index.html
@@ -10,8 +10,6 @@
     <script src="node_modules/blockly/blocks_compressed.js"></script>
     <script src="node_modules/blockly/msg/js/en.js"></script>
     <script src="node_modules/blockly/javascript_compressed.js"></script>
-    <script src="lib/blocks.js"></script>
-    <script src="lib/javascript-commands.js"></script>
     <script src="bundle.js"></script>
   </body>
 </html>
