+++ src/blockly/index.js
@@ -1,0 +1,2 @@
+require('./blocks')
+require('./javascript-commands')
