+++ src/components/blockly.js
@@ -2,7 +2,7 @@
 const widget = require('cache-element/widget')
 const html = require('choo/html')
 const sf = require('sheetify')
-const toolbox = require('../../lib/toolbox')
+const toolbox = require('../blockly/toolbox')
 
 const prefix = sf`
   :host, :host > .editor {
